done back end

+++ admin/controller.php
@@ -3,19 +3,31 @@
 	include("includes/sqlconnection.php");
 
 	// DELETE TO ANY TABLE
-	if (isset($_GET['id']) && isset($_GET['table'])) {
+	if (isset($_GET['id']) && isset($_GET['table']) && isset($_GET['control'])) {
 		
 		$id = $_GET['id'];
 		$table = $_GET['table'];
+		$control = $_GET['control'];
 		
 		$conn->query("DELETE FROM $table WHERE id='$id'");
 	
-		$_SESSION['status'] = "Record Deleted Successfully";
-		header('location:admin.php');
+
+		if($table == 'prodtable' || $table == 'cart' && $control == 'false'){
+			$_SESSION['status'] = "Record Deleted Successfully";
+			header('location:admin.php');
+		}
+		else if ($table == 'cart' && $control == 'true'){
+			$_SESSION['status'] = "Record Deleted Successfully";
+			header('location:../cart.php');
+		}
+
+
+
 		$conn->close();
 	} else {
 		echo "ID or table parameter missing";
 	}
+
 
 
 	// INSERT TO PRODUCT TABLE
@@ -82,183 +94,100 @@
 		$conn->close();
 	}	
 		
-
-
-	// INSERT TO NAVIGATION TABLE
-	if (isset($_POST['save_nav'])) {
+	// INSERT TO CART TABLE FROM ADD TO CART BUTTON
+	if (isset($_POST['save_cart'])) {
 		
+		$prodid = $_POST['txtprodid'];
+		$prodname = $_POST['txtprodname'];
+		$proddesc = $_POST['txtproddesc'];
+		$price = $_POST['txtprice'];
+		$image = $_FILES['txtimage']['name'];
+		$category = $_POST['txtcategory'];
 
-		$brand = $_POST['txtBrand'];
-		$nav1 = $_POST['txtNav1'];
-		$nav2 = $_POST['txtNav2'];
-		$nav3 = $_POST['txtNav3'];
-		$nav4 = $_POST['txtNav4'];
+		move_uploaded_file($_FILES['txtimage']['tmp_name'], "uploads/" . $image);
 
-
-
-		$sql_content = "INSERT INTO navbar(brand, nav1, nav2, nav3, nav4)
-		VALUES ('$brand', '$nav1', '$nav2', '$nav3', '$nav4')";
+		$sql_content = "INSERT INTO cart(prodid, prodname, proddesc, price, image, category)
+		VALUES ('$prodid', '$prodname', '$proddesc' , '$price', '$image', '$category' )";
 
 		if ($conn->query($sql_content) === TRUE) {
 			$id = $conn->insert_id;
 
-			$_SESSION['status'] = "Portfolio Content Record Inserted Successfully";
-			header('location:admin.php');
+			$_SESSION['status'] = "Success";
+			header($location);
 		}
 		else {
-			$_SESSION['status'] = "Error: Insert Failed.....";
-			header('location:insert_nav.php');
+			$_SESSION['status'] = "Failed";
+			header($location);
 		}
 		$conn->close();
 	}
 
-	// UPDATE TO NAVIGATION TABLE
-	if(isset($_POST['update_nav'])){
+
+		// UPDATE TO CART TABLE
+		if(isset($_POST['update_cart'])){
+
+			$id = $_POST['txtid'];
+			$prodid = $_POST['txtprodid'];
+			$prodname = $_POST['txtprodname'];
+			$proddesc = $_POST['txtproddesc'];
+			$price = $_POST['txtprice'];
+			$image_new = $_FILES['txtimage']['name'];
+			$image_old = $_POST['txtold_image'];
+			$category = $_POST['txtcategory'];
 	
-		$id = $_POST['txtid'];
-		$brand = $_POST['txtBrand'];
-		$nav1 = $_POST['txtNav1'];
-		$nav2 = $_POST['txtNav2'];
-		$nav3 = $_POST['txtNav3'];
-		$nav4 = $_POST['txtNav4'];
-
-		$sql_content = "UPDATE navbar SET brand ='$brand', nav1 = '$nav1', nav2  = '$nav2', nav3 = '$nav3',  nav4 = '$nav4' WHERE id='$id'";
 	
-		if ($conn->query($sql_content) === TRUE) {
-			$_SESSION['status'] = "Portfolio Content Record Updated Successfully";
-			header('location:admin.php');
-		} else {
-			$_SESSION['status'] = "Error: Update Failed";
-			header('location:edit_nav.php?id=' . $id);
-		}
-		
-		$conn->close();
-	}
-
-
-	// INSERT TO SECOND SECTION TABLE
-	if (isset($_POST['save_secondSection'])) {
-		
-		$image = $_FILES['txtImage']['name'];
-
-		move_uploaded_file($_FILES['txtImage']['tmp_name'], "uploads/" . $image);
-
-		$sql_content = "INSERT INTO secondSection(image)
-		VALUES ('$image')";
-
-		if($conn->query($sql_content) === TRUE){
-			$id = $conn->insert_id;
-		
-			$_SESSION['status'] = "Image Sucessfully Uploaded";
-			header('location:admin.php');
-			
-		}else{
-			
-			$_SESSION['status'] = "Error: Insert Failed.....";
-			header('location:insert_secondSection.php');
-
-		}
-		$conn->close();
-	}	
-
-	// UPDATE TO SECOND SECTION TABLE
-	if(isset($_POST['update_secondSection'])){
+			if($image_new != ''){
+				$update_image = $image_new;
+				
+			}else{
+				$update_image = $image_old;
+			}
 	
-		$id = $_POST['txtid'];
-		$image_new = $_FILES['txtImage']['name'];
-		$image_old = $_POST['txtold_image'];
-
+			echo "$update_image";
 	
-		if($image_new != ''){
-			$update_image = $image_new;
-			
-		}else{
-			$update_image = $image_old;
-		}
-
-		echo "$update_image";
-
-		move_uploaded_file($_FILES['txtImage']['tmp_name'], "uploads/" . $update_image);
-
-		$sql_content = "UPDATE secondSection SET image = '$update_image' WHERE id='$id'";
-		
-		if($conn->query($sql_content) === TRUE){
-			$id = $conn->insert_id;
-		
-			$_SESSION['status'] = "Image Sucessfully Uploaded";
-			header('location:admin.php');
-			
-		}else{
-			
-			$_SESSION['status'] = "Error: Insert Failed.....";
-			header('location:insert_secondSection.php');
-
-		}
-
-		$conn->close();
-	}	
-
-	// INSERT TO THIRD SECTION TABLE
-	if (isset($_POST['save_thirdSection'])) {
-		
-		$profile = $_FILES['txtProfile']['name'];
-		$description = $_POST['txtDescription'];
-
-		move_uploaded_file($_FILES['txtProfile']['tmp_name'], "uploads/" . $profile);
-
-		$sql_content = "INSERT INTO thirdSection(profile, description)
-		VALUES ('$profile', 'description')";
-
-		if($conn->query($sql_content) === TRUE){
-			$id = $conn->insert_id;
-		
-			$_SESSION['status'] = "Image Sucessfully Uploaded";
-			header('location:admin.php');
-			
-		}else{
-			
-			$_SESSION['status'] = "Error: Insert Failed.....";
-			header('location:insert_secondSection.php');
-
-		}
-		$conn->close();
-	}
-
-	// UPDATE TO THIRD SECTION TABLE
-	if(isset($_POST['update_thirdSection'])){
+			move_uploaded_file($_FILES['txtimage']['tmp_name'], "uploads/" . $update_image);
 	
-		$id = $_POST['txtid'];
-		$profile_new = $_FILES['txtProfile']['name'];
-		$profile_old = $_POST['txtold_profile'];
-		$description = $_POST['txtDescription'];
+			$sql_content = "UPDATE cart SET prodid = '$prodid', prodname = '$prodname', proddesc = '$proddesc', price = '$price', image = '$update_image', category = '$category'  WHERE id='$id'";
+		
+			if ($conn->query($sql_content) === TRUE) {
+				$_SESSION['status'] = "Success";
+				header('location:admin.php');
+			} else {
+				$_SESSION['status'] = "Failed";
+				header('location:edit_cart.php?id=' . $id);
+			}
+			
+			$conn->close();
+		}	
 
+
+			// INSERT TO CART TABLE
+		if (isset($_POST['save_cartprod'])) {
+			
+			$prodid = $_POST['txtprodid'];
+			$prodname = $_POST['txtprodname'];
+			$proddesc = $_POST['txtproddesc'];
+			$price = $_POST['txtprice'];
+			$image = $_POST['txtimage'];
+			$category = $_POST['txtcategory'];
+
+
+			$sql_content = "INSERT INTO cart(prodid, prodname, proddesc, price, image, category)
+			VALUES ('$prodid', '$prodname', '$proddesc' , '$price', '$image', '$category' )";
+
+			$location = 'location:../product.php?id=' . $prodid . '';
+
+			if ($conn->query($sql_content) === TRUE) {
+				$id = $conn->insert_id;
+
+				$_SESSION['status'] = "Success";
+				header($location);
+			}
+			else {
+				$_SESSION['status'] = "Failed";
+				header($location);
+			}
+			$conn->close();
+		}
 	
-		if($profile_new != ''){
-			$profile_image = $profile_new;
-			
-		}else{
-			$profile_image = $profile_old;
-		}
-
-		echo "$profile_image";
-
-		move_uploaded_file($_FILES['txtProfile']['tmp_name'], "uploads/" . $profile_image);
-
-		$sql_content = "UPDATE thirdSection SET profile = '$profile_image', description = '$description' WHERE id='$id'";
-		
-		if($conn->query($sql_content) === TRUE){
-			$id = $conn->insert_id;
-		
-			$_SESSION['status'] = "Image Sucessfully Uploaded";
-			header('location:admin.php');
-			
-		}else{
-			
-			$_SESSION['status'] = "Error: Insert Failed.....";
-			header('location:insert_thirdSection.php');
-
-		}
-
-		$conn->close();
-	}	
 ?>
